# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The backend powering most things",
        "title": "web-api",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ystv/web-api",
            "email": "computing@ystv.co.uk"
        },
        "license": {}
    },
    "paths": {
...
        "/v1/internal/creator/calendar/{year}/{month}": {
            "get": {
                "description": "Lists videos by month.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator-videos"
                ],
                "summary": "List videos by month",
                "operationId": "get-creator-videos-calendar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "year",
                        "name": "year",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "month",
                        "name": "month",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/video.MetaCal"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/internal/creator/videos": {
            "get": {
                "description": "Lists all videos, doesn't include files inside.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator-videos"
                ],
                "summary": "List all videos",
                "operationId": "get-creator-videos-all",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/video.Meta"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "creates a new video, requires the file ID/name to find it in CDN.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "creator-videos"
                ],
                "summary": "New video",
                "operationId": "new-creator-video",
                "parameters": [
                    {
                        "description": "NewVideo object",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/video.New"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Video ID",
                        "schema": {
                            "type": "body"
                        }
                    }
                }
            }
        },
...
        "/v1/internal/creator/videos/{videoid}": {
            "get": {
                "description": "Get a playlist including it's children files.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "creator-videos"
                ],
                "summary": "Get video by ID",
                "operationId": "get-creator-video",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Video ID",
                        "name": "videoid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/video.Item"
                        }
                    }
                }
            }
        },
        "/v1/internal/misc/list/{listid}": {
            "get": {
                "description": "Get a mailing list. Provides list subscribers also",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "misc-list"
                ],
                "summary": "Get Mailing list",
                "operationId": "get-mailing-list-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.List"
                        }
                    }
                }
            }
        },
        "/v1/internal/misc/list/{listid}/subscribe": {
            "post": {
                "description": "Subscribe to a mailing list by a JWT",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "misc-list"
                ],
                "summary": "Subscribe to mailing list by token",
                "operationId": "new-mailing-list-subscriber-token",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {}
                }
            }
        },
        "/v1/internal/misc/list/{listid}/subscribe/{userid}": {
            "post": {
                "description": "Subscribe to a mailing list by a user ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "misc-list"
                ],
                "summary": "Subscribe to mailing list by user ID",
                "operationId": "new-mailing-list-subscriber-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {}
                }
            }
        },
        "/v1/internal/misc/list/{listid}/subscribers": {
            "get": {
                "description": "Get a mailing list's subscribers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "misc-list"
                ],
                "summary": "Get subscribers",
                "operationId": "get-mailing-list-subscribers-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.List"
                        }
                    }
                }
            }
        },
        "/v1/internal/misc/list/{listid}/unsubscribe": {
            "delete": {
                "description": "Unsubscribe to a mailing list by a JWT",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "misc-list"
                ],
                "summary": "Unsubscribe to mailing list by token",
                "operationId": "delete-mailing-list-subscriber-token",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
        "/v1/internal/misc/list/{listid}/unsubscribe/{userid}": {
            "delete": {
                "description": "Unsubscribe to a mailing list by a user ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "misc-list"
                ],
                "summary": "Unsubscribe to mailing list by user ID",
                "operationId": "delete-mailing-list-subscriber-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "listid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
...
        "/v1/list_unsubscribe/{uuid}": {
            "get": {
                "description": "Unsubscribe to a mailing list by a subscriber UUID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "misc-list"
                ],
                "summary": "Unsubscribe to mailing list by subscriber UUID",
                "operationId": "delete-mailing-list-subscriber-uuid",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Subscriber UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
...
    },
    "definitions": {
...
        "misc.List": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "isSubscribed": {
                    "type": "boolean"
                },
                "listID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "permissionID": {
                    "type": "string"
                },
                "subscribers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/misc.Subscriber"
                    }
                }
            }
        },
...
        "misc.Subscriber": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/people.Permission"
                    }
                },
                "subscribeID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "people.Permission": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "video.File": {
            "type": "object",
            "properties": {
                "encodeFormat": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "video.Item": {
            "type": "object",
            "properties": {
                "broadcastDate": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/video.File"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "presetID": {
                    "type": "string"
                },
                "seriesID": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                },
                "userNickname": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "video.Meta": {
            "type": "object",
            "properties": {
                "broadcastDate": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "presetID": {
                    "type": "string"
                },
                "seriesID": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                },
                "userNickname": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "video.MetaCal": {
            "type": "object",
            "properties": {
                "broadcastDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "video.New": {
            "type": "object",
            "properties": {
                "broadcastDate": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "fileID": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "preset": {
                    "type": "integer"
                },
                "publishType": {
                    "type": "string"
                },
                "seriesID": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "urlName": {
                    "type": "string"
                }
            }
        }
    }
}